refactor: extract GenreFilter component to new file
