+++ src/views/components/GenreFilter/IGenreFilter.ts
@@ -1,0 +1,7 @@
+import { ComponentProps } from 'react'
+import { Genre } from '@data/types'
+
+
+export interface IGenreFilter extends ComponentProps<'menu'> {
+	genres: Genre[]
+}
