+++ src/views/components/Selection/Selection.tsx
@@ -1,9 +1,9 @@
-import { Genre } from '@data/types'
 import { ISelection } from './ISelection'
 import style from './Selection.module.scss'
-import { ComponentProps, useState } from 'react'
+import { useState } from 'react'
 import { GamePageList } from '@views/components'
 import { SelectionViewModel } from '@src/view-models/SelectionViewModel.ts'
+import { GenreFilter } from '@views/components/GenreFilter/GenreFilter.tsx'
 
 
 const viewModel = new SelectionViewModel()
@@ -23,24 +23,3 @@ export function Selection({ games, genres }: ISelection) {
 		</article>)
 }
 
-interface IGenreFilter extends ComponentProps<'menu'> {
-	genres: Genre[]
-}
-
-function GenreFilter({ genres }: IGenreFilter) {
-	const genreList = genres.map(g => {
-		const key = `selection-${typeof (g)}-${g.id}`
-
-		return (
-			<li key={key} data-slug={g.slug}>
-				<button className={style.Option}>{g.name}</button>
-			</li>
-		)
-	})
-
-	return (
-		<menu className={style.Menu}>
-			{genreList}
-		</menu>
-	)
-}
