+++ src/views/components/index.ts
@@ -8,3 +8,4 @@ export * from './Gallery/Gallery'
 export * from './Selection/Selection'
 export * from './SearchFilter/SearchFilter'
 export * from './GamePageList/GamePageList'
+export { GenreFilter } from '@views/components/GenreFilter/GenreFilter.tsx'
